Force reload of module from debug script
The file Invoke-MetasysMethod.ps1 is really just for debugging purposes.
The problem is that modules are cached and so any change to the actual
module is not reflected in the next running of this script.

This change force the previously loaded module to be unloaded and the
new version loaded.