fix: preserve order in json properities where possible
A recent change in deserialization of JSON introduced a bad side effect.
The change parses all JSON using `-AsHashTable`. The problem is that if
one were to serialize that hashtable the resulting order of property names
is no longer guaranteed.

So this commit modifies the code so that at first when we call
`ConvertFrom-Json` we don't specify `-AsHashTable`. If an exception is thrown
(which will happen when parsing `/enumerations`) then we parse using
`-AsHashTable`. Now if a client serialize the object returned, they'll get
back JSON that preserves the ordering of property names in the original JSON.