code: add support for version to change on each request
In previous versions, once Login happened, the Version was fixed in stone.
No future request for the same session could use a different version of the
API. This made testing propblematic in the cases when one version of the API
was broken and a user wanted to fallback to an older version. This commit
fixes the problem.