feat: fail and write error on version conflict
It's possible to use an absolute url that has a version that conflicts
with the explicit Version parameter. If that happens error out and
write an error message.

I added this because during CR verifies I was using absolute URLs but
trying to test with an older version using Version flag. Didn't
notice my error right away and was confused by the incorrect payloads
I was seeing. This will prevent those errors.