docs: udpate readme with quick start
The previous README seemed to discuss how to use the built-in
PowerShell cmdlets to call services. Now it describes how to
install and use Invoke-MetasysMethod.

The QuickStart tries to stick just to the basics. Other docs,
tutorials and/or tips/trics are considered for the future to
provide more in-depth documentation.