fix: correct output and return types of functions
Fix #15

I didn't understand the output stream and what ends up in the output
stream. My use of `Write-Output` was a leading cause in ensuring that
`Invoke-MetasysModule` and other functions return an array of objects
rather than a `String`, `PSCustomObject` or `Hashtable` which is the
normal expected response.

This commit removes all uses of Write-Output. This is done by

* Ensuring each happy path thru a function emits just one object
  (string in most but not all cases) to the output stream.
* Using Write-Error for some cases related to errors
* Using Write-Information for some cases