Only change background colors when running in console mode
I commited this change weeks after finding the bug. Basically,
IIRC when any of these scripts are run in a "headless" fashion
(say a cron job??) then there is no console for which to adjust
background colors. So the code that attempts to change background
colors throws a bunch of errors. This resolves that issue.