feat: add config file support
See README for more details.

Connect-MetasysAccount will now consult a file named
`$HOME/.metasysrestclient`. This is expected to be JSON file with a
property named `hosts`. This is an array of "host entries". Each host
entry must needs at a minimum an "alias" and a "hostname" property. This
allows you to specify an alias for -MetasysHost when calling
Connect-MetasysAccount.

Other properties that can be in a host entry are

- username
- version
- skip-certificate-check