bug: sort headers in tests before checking for equality
There was so much variability over time in the order that
Headers on response bodies would be serialized for checking
actuals/expecteds in unit tests.

Now we create a response body with the headers sorted (from the
actual response) and compare that to an expected response with
headers already sorted.