test: resolve long delay in test run
I don't really understand the root cause. But a BeforeAll was inside of
an `It` instead of outside of it. Perhaps this caused it to be called
multiple times? It's not clear. But this is definitely the test that
slowed everything down. With this commit in place tests run in a few seconeds
versus a minute or so.