test: user preferences for defaults are used in cma
This checks to make sure that if the respective parameters are not passed to
`cma`, then the user preferences for version and skip cert check are used.

The preferences are determined by the values of

- $env:METASYS_DEFAULT_API_VERSION
- $env:METASYS_SKIP_CHECK_NOT_SECURE

So if version isn't specified in a call to cma, then the default is used.
And if SkipCertificateCheck isn't specified in a call to cma, then the default is used.