bug: use skip-check from config file if set
A recent change to honor the default user preference for skip certificate check,
broke the code that tries to use the host entry information associated with an alias.

What was happening is that the code was checking to see if SkipCertificateCheck was
specified on the cli and if not then returning the value from
$env:METASYS_SKIP_CHECK_NOT_SECURE which if not set will return $false.

Now the code checks to see if the switch is not already set before falling back
to the user preference.