bug: fix issue with line breaks when running on windows
Two of the tests are comparing actual to expected http responses
in string format. To account for serialization variability in
headers we sort them before comparing. But when putting the responses
back together we end up with newlines that match the environment it is
run on. But this didn't match the expecation which had linux line
endings

Long story short: we now run actual and expected thru a little
dos2unix conversion to normalize the line endings.