bug: resolve issue with secure strings in tests
The tests that check the $env:METASYS_ACESS_TOKEN was set properly
failed on Windows. It appears that one or both of the
ConvertTo-SecureString/ConvertFrom-SecureString methods uses an encoding
that is time dependent. So you can get to differently encoded strings
for the same input. To resolve this, I just decode the access token back
into plain text and compare it to the expecte plain text.