fix: correct response when it is plain text
There were some scenarios where a plain text file was returned but the
Content-Length header was not set. The internal logic just defaults the
Content-Type to "json" when there is no Content-Length. This commit
fixes this bug.

The symptom of this bug is that the plain text file would be embedded within
another string. Meaning the output of the call would have double quotes around
it and all line feeds in the content would be escaped. This is now resolved.