fix: use temp files for saving responses
This program used to save the responses from `imm` in
`$env:METASYS_LAST_RESPONSE`.
But in general responses can be large and session errors can happen
when the total length of env variable characters is greater than 32,767.

So this commit now writes the last result to a temp file and stores the path
to that temp file in $env:METASYS_LAST_RESPONSE_PATH